Component-Video-Carousel -> Ajustments and realocated component to components folder
